requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
sqlite3

# AI and Media Processing
openai==1.3.7
whisper==1.1.10
opencv-python==4.8.1.78
pillow==10.1.0
pytesseract==0.3.10
requests==2.31.0
youtube-dl==2021.12.17
yt-dlp==2023.11.16

# Image and Video Processing
imageio==2.33.0
imageio-ffmpeg==0.4.9
numpy==1.24.3
scikit-image==0.21.0

# Text Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Utilities
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1 
httpx==0.27.0 